spline #108 refactoring step 9: resolve last fixme + plugin ordering
