+++ core/src/main/scala/za/co/absa/spline/harvester/LineageHarvesterFactory.scala
@@ -19,7 +19,7 @@ package za.co.absa.spline.harvester
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.execution.SparkPlan
-import za.co.absa.spline.harvester.builder.PluggableDataSourceFormatResolver
+import za.co.absa.spline.harvester.builder.dsformat.PluggableDataSourceFormatResolver
 import za.co.absa.spline.harvester.builder.read.PluggableReadCommandExtractor
 import za.co.absa.spline.harvester.builder.write.PluggableWriteCommandExtractor
 import za.co.absa.spline.harvester.conf.SplineConfigurer.SplineMode.SplineMode
