+++ core/src/main/scala/za/co/absa/spline/harvester/plugin/composite/SaveIntoDataSourceCommandPlugin.scala
@@ -16,16 +16,18 @@
 
 package za.co.absa.spline.harvester.plugin.composite
 
+import javax.annotation.Priority
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.execution.datasources.SaveIntoDataSourceCommand
 import za.co.absa.commons.reflect.extractors.AccessorMethodValueExtractor
 import za.co.absa.spline.harvester.builder.SourceIdentifier
-import za.co.absa.spline.harvester.plugin.Plugin.WriteNodeInfo
+import za.co.absa.spline.harvester.plugin.Plugin.{Precedence, WriteNodeInfo}
 import za.co.absa.spline.harvester.plugin.composite.SaveIntoDataSourceCommandPlugin._
 import za.co.absa.spline.harvester.plugin.registry.PluginRegistry
 import za.co.absa.spline.harvester.plugin.{Plugin, RelationProviderProcessing, WriteNodeProcessing}
 import za.co.absa.spline.harvester.qualifier.PathQualifier
 
+@Priority(Precedence.Lowest)
 class SaveIntoDataSourceCommandPlugin(
   pluginRegistry: PluginRegistry,
   pathQualifier: PathQualifier)
@@ -39,20 +41,18 @@ class SaveIntoDataSourceCommandPlugin(
 
 
   override def writeNodeProcessor: PartialFunction[LogicalPlan, WriteNodeInfo] = {
-    // fixme: should the default case be handled here?
-    case cmd: SaveIntoDataSourceCommand =>
-      cmd match {
-        case RelationProviderExtractor(rp)
-          if rpProcessor.isDefinedAt((rp, cmd)) =>
-          rpProcessor((rp, cmd))
+    case cmd: SaveIntoDataSourceCommand => cmd match {
+      case RelationProviderExtractor(rp)
+        if rpProcessor.isDefinedAt((rp, cmd)) =>
+        rpProcessor((rp, cmd))
 
-        case _ =>
-          val maybeProvider = RelationProviderExtractor.unapply(cmd)
-          val opts = cmd.options
-          val uri = opts.get("path").map(pathQualifier.qualify)
-            .getOrElse(sys.error(s"Cannot extract source URI from the options: ${opts.keySet mkString ","}"))
-          (SourceIdentifier(maybeProvider, uri), cmd.mode, cmd.query, opts)
-      }
+      case _ =>
+        val maybeProvider = RelationProviderExtractor.unapply(cmd)
+        val opts = cmd.options
+        val uri = opts.get("path").map(pathQualifier.qualify)
+          .getOrElse(sys.error(s"Cannot extract source URI from the options: ${opts.keySet mkString ","}"))
+        (SourceIdentifier(maybeProvider, uri), cmd.mode, cmd.query, opts)
+    }
   }
 }
