+++ core/src/main/scala/za/co/absa/spline/harvester/plugin/embedded/CobrixPlugin.scala
@@ -16,16 +16,18 @@
 
 package za.co.absa.spline.harvester.plugin.embedded
 
+import javax.annotation.Priority
 import org.apache.spark.sql.execution.datasources.LogicalRelation
 import org.apache.spark.sql.sources.BaseRelation
 import za.co.absa.commons.reflect.ReflectionUtils.extractFieldValue
 import za.co.absa.commons.reflect.extractors.SafeTypeMatchingExtractor
 import za.co.absa.spline.harvester.builder.SourceIdentifier
-import za.co.absa.spline.harvester.plugin.Plugin.ReadNodeInfo
+import za.co.absa.spline.harvester.plugin.Plugin.{Precedence, ReadNodeInfo}
 import za.co.absa.spline.harvester.plugin.embedded.CobrixPlugin._
 import za.co.absa.spline.harvester.plugin.{BaseRelationProcessing, Plugin}
 
 
+@Priority(Precedence.Normal)
 class CobrixPlugin extends Plugin with BaseRelationProcessing {
 
   override def baseRelationProcessor: PartialFunction[(BaseRelation, LogicalRelation), ReadNodeInfo] = {
