+++ core/src/main/scala/za/co/absa/spline/harvester/plugin/embedded/KafkaPlugin.scala
@@ -18,6 +18,7 @@ package za.co.absa.spline.harvester.plugin.embedded
 
 import java.util.Properties
 
+import javax.annotation.Priority
 import org.apache.kafka.clients.consumer.KafkaConsumer
 import org.apache.spark.sql.execution.datasources.{LogicalRelation, SaveIntoDataSourceCommand}
 import org.apache.spark.sql.kafka010.{AssignStrategy, ConsumerStrategy, SubscribePatternStrategy, SubscribeStrategy}
@@ -25,12 +26,13 @@ import org.apache.spark.sql.sources.BaseRelation
 import za.co.absa.commons.reflect.ReflectionUtils.extractFieldValue
 import za.co.absa.commons.reflect.extractors.SafeTypeMatchingExtractor
 import za.co.absa.spline.harvester.builder.SourceIdentifier
-import za.co.absa.spline.harvester.plugin.Plugin.{ReadNodeInfo, WriteNodeInfo}
+import za.co.absa.spline.harvester.plugin.Plugin.{Precedence, ReadNodeInfo, WriteNodeInfo}
 import za.co.absa.spline.harvester.plugin.embedded.KafkaPlugin._
-import za.co.absa.spline.harvester.plugin.{BaseRelationProcessing, RelationProviderProcessing, Plugin}
+import za.co.absa.spline.harvester.plugin.{BaseRelationProcessing, Plugin, RelationProviderProcessing}
 
 import scala.collection.JavaConverters._
 
+@Priority(Precedence.Normal)
 class KafkaPlugin
   extends Plugin
     with BaseRelationProcessing
